rewrote reader, started with saver

+++ Modbus.Core/IModbus.cs
@@ -34,12 +34,12 @@ namespace Modbus.Core
         ModbusErrorCode PresetMultipleRegisters(Byte rtuAddress, UInt16 forceAddress, List<object> values, bool bigEndianOrder = false);
 
         //overloads with indexing in object array
-        ModbusErrorCode ReadHoldingRegisters(Byte rtuAddress, UInt16 startAddress, ref object[] registerValues, UInt16 startIndex, UInt16 objectsCount, bool bigEndianOrder = false);
-        ModbusErrorCode ReadInputRegisters(Byte rtuAddress, UInt16 startAddress, ref object[] registerValues, UInt16 startIndex, UInt16 objectsCount, bool bigEndianOrder = false);
-        ModbusErrorCode ReadHoldingRegisters(Byte rtuAddress, UInt16 startAddress, ref List<object> registerValues, UInt16 startIndex, UInt16 objectsCount, bool bigEndianOrder = false);
-        ModbusErrorCode ReadInputRegisters(Byte rtuAddress, UInt16 startAddress, ref List<object> registerValues, UInt16 startIndex, UInt16 objectsCount, bool bigEndianOrder = false);
-        ModbusErrorCode PresetMultipleRegisters(Byte rtuAddress, UInt16 forceAddress, object[] values, UInt16 startIndex, UInt16 objectsCount, bool bigEndianOrder = false);
-        ModbusErrorCode PresetMultipleRegisters(Byte rtuAddress, UInt16 forceAddress, List<object> values, UInt16 startIndex, UInt16 objectsCount, bool bigEndianOrder = false);
+        ModbusErrorCode ReadHoldingRegisters(Byte rtuAddress, UInt16 startAddress, ref object[] registerValues, UInt32 startIndex, UInt32 objectsCount, bool bigEndianOrder = false);
+        ModbusErrorCode ReadInputRegisters(Byte rtuAddress, UInt16 startAddress, ref object[] registerValues, UInt32 startIndex, UInt32 objectsCount, bool bigEndianOrder = false);
+        ModbusErrorCode ReadHoldingRegisters(Byte rtuAddress, UInt16 startAddress, ref List<object> registerValues, UInt32 startIndex, UInt32 objectsCount, bool bigEndianOrder = false);
+        ModbusErrorCode ReadInputRegisters(Byte rtuAddress, UInt16 startAddress, ref List<object> registerValues, UInt32 startIndex, UInt32 objectsCount, bool bigEndianOrder = false);
+        ModbusErrorCode PresetMultipleRegisters(Byte rtuAddress, UInt16 forceAddress, object[] values, UInt32 startIndex, UInt32 objectsCount, bool bigEndianOrder = false);
+        ModbusErrorCode PresetMultipleRegisters(Byte rtuAddress, UInt16 forceAddress, List<object> values, UInt32 startIndex, UInt32 objectsCount, bool bigEndianOrder = false);
         bool IsConnected { get; }
         Byte ReadRegistersPerQueryCapacity { get; set; }
         Byte WriteRegistersPerQueryCapacity { get; set; }
